refactor: rename QuestionListSection to QuestionList for clarity
- Rename QuestionListSection.tsx to QuestionList.tsx to better reflect its purpose
- Update imports in NurseryDetailPage.tsx to use new component name
- Fix QuestionList.test.tsx to match new component interface
- Remove outdated component and update related test files
